isolate Lion from Feline improve AnimalTest and LionTest with Enclosed
